Removed GameObject Reference from all Transforms
This was a bad idea to begin with. This complicated the logic of the
code and the syntax (because of forward declarations).
This affected the TransparentTransform because it got the camera postion
from the GameObject. This will need to be fixed in the future.
GameObjectUtils and GameObjectResourceManager need cleaning up because
they still take reference to GameObject but don't actually use the
reference.